[SPARK-46278][CONNECT][PS][TESTS] Re-organize `GroupByTests`
### What changes were proposed in this pull request?

This PR proposes to re-organize `GroupByTests`.

### Why are the changes needed?

Break the big test file into small groups by topic such as Pandas does.

### Does this PR introduce _any_ user-facing change?

No, it's test-only.

### How was this patch tested?

The existing CI should pass with new structure of test files.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44196 from itholic/reorg_groupby_test.

Authored-by: Haejoon Lee <haejoon.lee@databricks.com>
