+++ dev/sparktestsupport/modules.py
@@ -815,11 +815,18 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.groupby.test_cumulative",
         "pyspark.pandas.tests.groupby.test_describe",
         "pyspark.pandas.tests.groupby.test_groupby",
+        "pyspark.pandas.tests.groupby.test_grouping",
         "pyspark.pandas.tests.groupby.test_head_tail",
         "pyspark.pandas.tests.groupby.test_index",
+        "pyspark.pandas.tests.groupby.test_missing",
         "pyspark.pandas.tests.groupby.test_missing_data",
+        "pyspark.pandas.tests.groupby.test_nlargest_nsmallest",
+        "pyspark.pandas.tests.groupby.test_raises",
+        "pyspark.pandas.tests.groupby.test_rank",
+        "pyspark.pandas.tests.groupby.test_size",
         "pyspark.pandas.tests.groupby.test_split_apply",
         "pyspark.pandas.tests.groupby.test_stat",
+        "pyspark.pandas.tests.groupby.test_value_counts",
         "pyspark.pandas.tests.test_indexing",
         "pyspark.pandas.tests.test_ops_on_diff_frames",
         "pyspark.pandas.tests.test_ops_on_diff_frames_groupby",
@@ -1113,6 +1120,13 @@ pyspark_pandas_connect_part2 = Module(
         "pyspark.pandas.tests.connect.groupby.test_parity_describe",
         "pyspark.pandas.tests.connect.groupby.test_parity_head_tail",
         "pyspark.pandas.tests.connect.groupby.test_parity_groupby",
+        "pyspark.pandas.tests.connect.groupby.test_parity_grouping",
+        "pyspark.pandas.tests.connect.groupby.test_parity_missing",
+        "pyspark.pandas.tests.connect.groupby.test_parity_nlargest_nsmallest",
+        "pyspark.pandas.tests.connect.groupby.test_parity_raises",
+        "pyspark.pandas.tests.connect.groupby.test_parity_rank",
+        "pyspark.pandas.tests.connect.groupby.test_parity_size",
+        "pyspark.pandas.tests.connect.groupby.test_parity_value_counts",
     ],
     excluded_python_implementations=[
         "PyPy"  # Skip these tests under PyPy since they require numpy, pandas, and pyarrow and
