+++ python/pyspark/pandas/tests/groupby/test_groupby.py
@@ -186,84 +186,6 @@ class GroupByTestsMixin:
                 sort(pdf.groupby(10, as_index=as_index)[[20, 30]].sum()),
             )
 
-    def test_nsmallest(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 1, 1, 2, 2, 2, 3, 3, 3] * 3,
-                "b": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-                "c": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-                "d": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-            },
-            index=np.random.rand(9 * 3),
-        )
-        psdf = ps.from_pandas(pdf)
-
-        self.assert_eq(
-            psdf.groupby(["a"])["b"].nsmallest(1).sort_values(),
-            pdf.groupby(["a"])["b"].nsmallest(1).sort_values(),
-        )
-        self.assert_eq(
-            psdf.groupby(["a"])["b"].nsmallest(2).sort_index(),
-            pdf.groupby(["a"])["b"].nsmallest(2).sort_index(),
-        )
-        self.assert_eq(
-            (psdf.b * 10).groupby(psdf.a).nsmallest(2).sort_index(),
-            (pdf.b * 10).groupby(pdf.a).nsmallest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.rename().groupby(psdf.a).nsmallest(2).sort_index(),
-            pdf.b.rename().groupby(pdf.a).nsmallest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.groupby(psdf.a.rename()).nsmallest(2).sort_index(),
-            pdf.b.groupby(pdf.a.rename()).nsmallest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.rename().groupby(psdf.a.rename()).nsmallest(2).sort_index(),
-            pdf.b.rename().groupby(pdf.a.rename()).nsmallest(2).sort_index(),
-        )
-        with self.assertRaisesRegex(ValueError, "nsmallest do not support multi-index now"):
-            psdf.set_index(["a", "b"]).groupby(["c"])["d"].nsmallest(1)
-
-    def test_nlargest(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 1, 1, 2, 2, 2, 3, 3, 3] * 3,
-                "b": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-                "c": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-                "d": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
-            },
-            index=np.random.rand(9 * 3),
-        )
-        psdf = ps.from_pandas(pdf)
-
-        self.assert_eq(
-            psdf.groupby(["a"])["b"].nlargest(1).sort_values(),
-            pdf.groupby(["a"])["b"].nlargest(1).sort_values(),
-        )
-        self.assert_eq(
-            psdf.groupby(["a"])["b"].nlargest(2).sort_index(),
-            pdf.groupby(["a"])["b"].nlargest(2).sort_index(),
-        )
-        self.assert_eq(
-            (psdf.b * 10).groupby(psdf.a).nlargest(2).sort_index(),
-            (pdf.b * 10).groupby(pdf.a).nlargest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.rename().groupby(psdf.a).nlargest(2).sort_index(),
-            pdf.b.rename().groupby(pdf.a).nlargest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.groupby(psdf.a.rename()).nlargest(2).sort_index(),
-            pdf.b.groupby(pdf.a.rename()).nlargest(2).sort_index(),
-        )
-        self.assert_eq(
-            psdf.b.rename().groupby(psdf.a.rename()).nlargest(2).sort_index(),
-            pdf.b.rename().groupby(pdf.a.rename()).nlargest(2).sort_index(),
-        )
-        with self.assertRaisesRegex(ValueError, "nlargest do not support multi-index now"):
-            psdf.set_index(["a", "b"]).groupby(["c"])["d"].nlargest(1)
-
     def test_shift(self):
         pdf = pd.DataFrame(
             {
@@ -331,238 +253,10 @@ class GroupByTestsMixin:
         #                pdf.groupby([('x', 'a'), ('x', 'b')]).shift(periods=-1,
         #                                                            fill_value=0).sort_index())
 
-    def test_missing(self):
-        psdf = ps.DataFrame({"a": [1, 2, 3, 4, 5, 6, 7, 8, 9]})
-
-        # DataFrameGroupBy functions
-        missing_functions = inspect.getmembers(
-            MissingPandasLikeDataFrameGroupBy, inspect.isfunction
-        )
-        unsupported_functions = [
-            name for (name, type_) in missing_functions if type_.__name__ == "unsupported_function"
-        ]
-        for name in unsupported_functions:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError,
-                "method.*GroupBy.*{}.*not implemented( yet\\.|\\. .+)".format(name),
-            ):
-                getattr(psdf.groupby("a"), name)()
-
-        deprecated_functions = [
-            name for (name, type_) in missing_functions if type_.__name__ == "deprecated_function"
-        ]
-        for name in deprecated_functions:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError, "method.*GroupBy.*{}.*is deprecated".format(name)
-            ):
-                getattr(psdf.groupby("a"), name)()
-
-        # SeriesGroupBy functions
-        missing_functions = inspect.getmembers(MissingPandasLikeSeriesGroupBy, inspect.isfunction)
-        unsupported_functions = [
-            name for (name, type_) in missing_functions if type_.__name__ == "unsupported_function"
-        ]
-        for name in unsupported_functions:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError,
-                "method.*GroupBy.*{}.*not implemented( yet\\.|\\. .+)".format(name),
-            ):
-                getattr(psdf.a.groupby(psdf.a), name)()
-
-        deprecated_functions = [
-            name for (name, type_) in missing_functions if type_.__name__ == "deprecated_function"
-        ]
-        for name in deprecated_functions:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError, "method.*GroupBy.*{}.*is deprecated".format(name)
-            ):
-                getattr(psdf.a.groupby(psdf.a), name)()
-
-        # DataFrameGroupBy properties
-        missing_properties = inspect.getmembers(
-            MissingPandasLikeDataFrameGroupBy, lambda o: isinstance(o, property)
-        )
-        unsupported_properties = [
-            name
-            for (name, type_) in missing_properties
-            if type_.fget.__name__ == "unsupported_property"
-        ]
-        for name in unsupported_properties:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError,
-                "property.*GroupBy.*{}.*not implemented( yet\\.|\\. .+)".format(name),
-            ):
-                getattr(psdf.groupby("a"), name)
-        deprecated_properties = [
-            name
-            for (name, type_) in missing_properties
-            if type_.fget.__name__ == "deprecated_property"
-        ]
-        for name in deprecated_properties:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError, "property.*GroupBy.*{}.*is deprecated".format(name)
-            ):
-                getattr(psdf.groupby("a"), name)
-
-        # SeriesGroupBy properties
-        missing_properties = inspect.getmembers(
-            MissingPandasLikeSeriesGroupBy, lambda o: isinstance(o, property)
-        )
-        unsupported_properties = [
-            name
-            for (name, type_) in missing_properties
-            if type_.fget.__name__ == "unsupported_property"
-        ]
-        for name in unsupported_properties:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError,
-                "property.*GroupBy.*{}.*not implemented( yet\\.|\\. .+)".format(name),
-            ):
-                getattr(psdf.a.groupby(psdf.a), name)
-        deprecated_properties = [
-            name
-            for (name, type_) in missing_properties
-            if type_.fget.__name__ == "deprecated_property"
-        ]
-        for name in deprecated_properties:
-            with self.assertRaisesRegex(
-                PandasNotImplementedError, "property.*GroupBy.*{}.*is deprecated".format(name)
-            ):
-                getattr(psdf.a.groupby(psdf.a), name)
-
     @staticmethod
     def test_is_multi_agg_with_relabel():
         assert is_multi_agg_with_relabel(a="max") is False
         assert is_multi_agg_with_relabel(a_min=("a", "max"), a_max=("a", "min")) is True
-
-    def test_get_group(self):
-        pdf = pd.DataFrame(
-            [
-                ("falcon", "bird", 389.0),
-                ("parrot", "bird", 24.0),
-                ("lion", "mammal", 80.5),
-                ("monkey", "mammal", np.nan),
-            ],
-            columns=["name", "class", "max_speed"],
-            index=[0, 2, 3, 1],
-        )
-        pdf.columns.name = "Koalas"
-        psdf = ps.from_pandas(pdf)
-
-        self.assert_eq(
-            psdf.groupby("class").get_group("bird"),
-            pdf.groupby("class").get_group("bird"),
-        )
-        self.assert_eq(
-            psdf.groupby("class")["name"].get_group("mammal"),
-            pdf.groupby("class")["name"].get_group("mammal"),
-        )
-        self.assert_eq(
-            psdf.groupby("class")[["name"]].get_group("mammal"),
-            pdf.groupby("class")[["name"]].get_group("mammal"),
-        )
-        self.assert_eq(
-            psdf.groupby(["class", "name"]).get_group(("mammal", "lion")),
-            pdf.groupby(["class", "name"]).get_group(("mammal", "lion")),
-        )
-        self.assert_eq(
-            psdf.groupby(["class", "name"])["max_speed"].get_group(("mammal", "lion")),
-            pdf.groupby(["class", "name"])["max_speed"].get_group(("mammal", "lion")),
-        )
-        self.assert_eq(
-            psdf.groupby(["class", "name"])[["max_speed"]].get_group(("mammal", "lion")),
-            pdf.groupby(["class", "name"])[["max_speed"]].get_group(("mammal", "lion")),
-        )
-        self.assert_eq(
-            (psdf.max_speed + 1).groupby(psdf["class"]).get_group("mammal"),
-            (pdf.max_speed + 1).groupby(pdf["class"]).get_group("mammal"),
-        )
-        self.assert_eq(
-            psdf.groupby("max_speed").get_group(80.5),
-            pdf.groupby("max_speed").get_group(80.5),
-        )
-
-        self.assertRaises(KeyError, lambda: psdf.groupby("class").get_group("fish"))
-        self.assertRaises(TypeError, lambda: psdf.groupby("class").get_group(["bird", "mammal"]))
-        self.assertRaises(KeyError, lambda: psdf.groupby("class")["name"].get_group("fish"))
-        self.assertRaises(
-            TypeError, lambda: psdf.groupby("class")["name"].get_group(["bird", "mammal"])
-        )
-        self.assertRaises(
-            KeyError, lambda: psdf.groupby(["class", "name"]).get_group(("lion", "mammal"))
-        )
-        self.assertRaises(ValueError, lambda: psdf.groupby(["class", "name"]).get_group(("lion",)))
-        self.assertRaises(
-            ValueError, lambda: psdf.groupby(["class", "name"]).get_group(("mammal",))
-        )
-        self.assertRaises(ValueError, lambda: psdf.groupby(["class", "name"]).get_group("mammal"))
-
-        # MultiIndex columns
-        pdf.columns = pd.MultiIndex.from_tuples([("A", "name"), ("B", "class"), ("C", "max_speed")])
-        pdf.columns.names = ["Hello", "Koalas"]
-        psdf = ps.from_pandas(pdf)
-        self.assert_eq(
-            psdf.groupby(("B", "class")).get_group("bird"),
-            pdf.groupby(("B", "class")).get_group("bird"),
-        )
-        self.assert_eq(
-            psdf.groupby(("B", "class"))[[("A", "name")]].get_group("mammal"),
-            pdf.groupby(("B", "class"))[[("A", "name")]].get_group("mammal"),
-        )
-        self.assert_eq(
-            psdf.groupby([("B", "class"), ("A", "name")]).get_group(("mammal", "lion")),
-            pdf.groupby([("B", "class"), ("A", "name")]).get_group(("mammal", "lion")),
-        )
-        self.assert_eq(
-            psdf.groupby([("B", "class"), ("A", "name")])[[("C", "max_speed")]].get_group(
-                ("mammal", "lion")
-            ),
-            pdf.groupby([("B", "class"), ("A", "name")])[[("C", "max_speed")]].get_group(
-                ("mammal", "lion")
-            ),
-        )
-        self.assert_eq(
-            (psdf[("C", "max_speed")] + 1).groupby(psdf[("B", "class")]).get_group("mammal"),
-            (pdf[("C", "max_speed")] + 1).groupby(pdf[("B", "class")]).get_group("mammal"),
-        )
-        self.assert_eq(
-            psdf.groupby(("C", "max_speed")).get_group(80.5),
-            pdf.groupby(("C", "max_speed")).get_group(80.5),
-        )
-
-        self.assertRaises(KeyError, lambda: psdf.groupby(("B", "class")).get_group("fish"))
-        self.assertRaises(
-            TypeError, lambda: psdf.groupby(("B", "class")).get_group(["bird", "mammal"])
-        )
-        self.assertRaises(
-            KeyError, lambda: psdf.groupby(("B", "class"))[("A", "name")].get_group("fish")
-        )
-        self.assertRaises(
-            KeyError,
-            lambda: psdf.groupby([("B", "class"), ("A", "name")]).get_group(("lion", "mammal")),
-        )
-        self.assertRaises(
-            ValueError,
-            lambda: psdf.groupby([("B", "class"), ("A", "name")]).get_group(("lion",)),
-        )
-        self.assertRaises(
-            ValueError, lambda: psdf.groupby([("B", "class"), ("A", "name")]).get_group(("mammal",))
-        )
-        self.assertRaises(
-            ValueError, lambda: psdf.groupby([("B", "class"), ("A", "name")]).get_group("mammal")
-        )
-
-    def test_getitem(self):
-        psdf = ps.DataFrame(
-            {
-                "a": [1, 1, 1, 1, 2, 2, 2, 3, 3, 3] * 3,
-                "b": [2, 3, 1, 4, 6, 9, 8, 10, 7, 5] * 3,
-                "c": [3, 5, 2, 5, 1, 2, 6, 4, 3, 6] * 3,
-            },
-            index=np.random.rand(10 * 3),
-        )
-
-        self.assertTrue(isinstance(psdf.groupby("a")["b"], SeriesGroupBy))
 
     def test_all_any(self):
         pdf = pd.DataFrame(
@@ -647,19 +341,6 @@ class GroupByTestsMixin:
             pdf.groupby("A").all(skipna=True).sort_index(),
         )
 
-    def test_raises(self):
-        psdf = ps.DataFrame(
-            {"a": [1, 2, 6, 4, 4, 6, 4, 3, 7], "b": [4, 2, 7, 3, 3, 1, 1, 1, 2]},
-            index=[0, 1, 3, 5, 6, 8, 9, 9, 9],
-        )
-        # test raises with incorrect key
-        self.assertRaises(ValueError, lambda: psdf.groupby([]))
-        self.assertRaises(KeyError, lambda: psdf.groupby("x"))
-        self.assertRaises(KeyError, lambda: psdf.groupby(["a", "x"]))
-        self.assertRaises(KeyError, lambda: psdf.groupby("a")["x"])
-        self.assertRaises(KeyError, lambda: psdf.groupby("a")["b", "x"])
-        self.assertRaises(KeyError, lambda: psdf.groupby("a")[["b", "x"]])
-
     def test_nunique(self):
         pdf = pd.DataFrame(
             {"a": [1, 1, 1, 1, 1, 0, 0, 0, 0, 0], "b": [2, 2, 2, 3, 3, 4, 4, 5, 5, 5]}
@@ -727,99 +408,6 @@ class GroupByTestsMixin:
                 for act, exp in zip(actual, expect):
                     self.assertTrue(sorted(act) == sorted(exp))
 
-    def test_value_counts(self):
-        pdf = pd.DataFrame(
-            {"A": [np.nan, 2, 2, 3, 3, 3], "B": [1, 1, 2, 3, 3, np.nan]}, columns=["A", "B"]
-        )
-        psdf = ps.from_pandas(pdf)
-        self.assert_eq(
-            psdf.groupby("A")["B"].value_counts().sort_index(),
-            pdf.groupby("A")["B"].value_counts().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby("A")["B"].value_counts(dropna=False).sort_index(),
-            pdf.groupby("A")["B"].value_counts(dropna=False).sort_index(),
-            almost=True,
-        )
-        self.assert_eq(
-            psdf.groupby("A", dropna=False)["B"].value_counts(dropna=False).sort_index(),
-            pdf.groupby("A", dropna=False)["B"].value_counts(dropna=False).sort_index(),
-            # Returns are the same considering values and types,
-            # disable check_exact to pass the assert_eq
-            check_exact=False,
-        )
-        self.assert_eq(
-            psdf.groupby("A")["B"].value_counts(sort=True, ascending=False).sort_index(),
-            pdf.groupby("A")["B"].value_counts(sort=True, ascending=False).sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby("A")["B"]
-            .value_counts(sort=True, ascending=False, dropna=False)
-            .sort_index(),
-            pdf.groupby("A")["B"]
-            .value_counts(sort=True, ascending=False, dropna=False)
-            .sort_index(),
-            almost=True,
-        )
-        self.assert_eq(
-            psdf.groupby("A")["B"]
-            .value_counts(sort=True, ascending=True, dropna=False)
-            .sort_index(),
-            pdf.groupby("A")["B"]
-            .value_counts(sort=True, ascending=True, dropna=False)
-            .sort_index(),
-            almost=True,
-        )
-        self.assert_eq(
-            psdf.B.rename().groupby(psdf.A).value_counts().sort_index(),
-            pdf.B.rename().groupby(pdf.A).value_counts().sort_index(),
-        )
-        self.assert_eq(
-            psdf.B.rename().groupby(psdf.A, dropna=False).value_counts().sort_index(),
-            pdf.B.rename().groupby(pdf.A, dropna=False).value_counts().sort_index(),
-            # Returns are the same considering values and types,
-            # disable check_exact to pass the assert_eq
-            check_exact=False,
-        )
-        self.assert_eq(
-            psdf.B.groupby(psdf.A.rename()).value_counts().sort_index(),
-            pdf.B.groupby(pdf.A.rename()).value_counts().sort_index(),
-        )
-        self.assert_eq(
-            psdf.B.rename().groupby(psdf.A.rename()).value_counts().sort_index(),
-            pdf.B.rename().groupby(pdf.A.rename()).value_counts().sort_index(),
-        )
-
-    def test_size(self):
-        pdf = pd.DataFrame({"A": [1, 2, 2, 3, 3, 3], "B": [1, 1, 2, 3, 3, 3]})
-        psdf = ps.from_pandas(pdf)
-        self.assert_eq(psdf.groupby("A").size().sort_index(), pdf.groupby("A").size().sort_index())
-        self.assert_eq(
-            psdf.groupby("A")["B"].size().sort_index(), pdf.groupby("A")["B"].size().sort_index()
-        )
-        self.assert_eq(
-            psdf.groupby("A")[["B"]].size().sort_index(),
-            pdf.groupby("A")[["B"]].size().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(["A", "B"]).size().sort_index(),
-            pdf.groupby(["A", "B"]).size().sort_index(),
-        )
-
-        # multi-index columns
-        columns = pd.MultiIndex.from_tuples([("X", "A"), ("Y", "B")])
-        pdf.columns = columns
-        psdf.columns = columns
-
-        self.assert_eq(
-            psdf.groupby(("X", "A")).size().sort_index(),
-            pdf.groupby(("X", "A")).size().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby([("X", "A"), ("Y", "B")]).size().sort_index(),
-            pdf.groupby([("X", "A"), ("Y", "B")]).size().sort_index(),
-        )
-
     def test_diff(self):
         pdf = pd.DataFrame(
             {
@@ -867,56 +455,6 @@ class GroupByTestsMixin:
         self.assert_eq(
             psdf.groupby([("x", "a"), ("x", "b")]).diff().sort_index(),
             pdf.groupby([("x", "a"), ("x", "b")]).diff().sort_index(),
-        )
-
-    def test_rank(self):
-        pdf = pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5, 6] * 3,
-                "b": [1, 1, 2, 3, 5, 8] * 3,
-                "c": [1, 4, 9, 16, 25, 36] * 3,
-            },
-            index=np.random.rand(6 * 3),
-        )
-        psdf = ps.from_pandas(pdf)
-
-        self.assert_eq(psdf.groupby("b").rank().sort_index(), pdf.groupby("b").rank().sort_index())
-        self.assert_eq(
-            psdf.groupby(["a", "b"]).rank().sort_index(),
-            pdf.groupby(["a", "b"]).rank().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(["b"])["a"].rank().sort_index(),
-            pdf.groupby(["b"])["a"].rank().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(["b"])[["a", "c"]].rank().sort_index(),
-            pdf.groupby(["b"])[["a", "c"]].rank().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(psdf.b // 5).rank().sort_index(),
-            pdf.groupby(pdf.b // 5).rank().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby(psdf.b // 5)["a"].rank().sort_index(),
-            pdf.groupby(pdf.b // 5)["a"].rank().sort_index(),
-        )
-
-        self.assert_eq(psdf.groupby("b").rank().sum(), pdf.groupby("b").rank().sum())
-        self.assert_eq(psdf.groupby(["b"])["a"].rank().sum(), pdf.groupby(["b"])["a"].rank().sum())
-
-        # multi-index columns
-        columns = pd.MultiIndex.from_tuples([("x", "a"), ("x", "b"), ("y", "c")])
-        pdf.columns = columns
-        psdf.columns = columns
-
-        self.assert_eq(
-            psdf.groupby(("x", "b")).rank().sort_index(),
-            pdf.groupby(("x", "b")).rank().sort_index(),
-        )
-        self.assert_eq(
-            psdf.groupby([("x", "a"), ("x", "b")]).rank().sort_index(),
-            pdf.groupby([("x", "a"), ("x", "b")]).rank().sort_index(),
         )
