+++ python/pyspark/pandas/tests/groupby/test_nlargest_nsmallest.py
@@ -1,0 +1,119 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import unittest
+
+import pandas as pd
+import numpy as np
+import pyspark.pandas as ps
+from pyspark.testing.pandasutils import PandasOnSparkTestCase
+
+
+class NlargestNsmallestTestsMixin:
+    def test_nlargest(self):
+        pdf = pd.DataFrame(
+            {
+                "a": [1, 1, 1, 2, 2, 2, 3, 3, 3] * 3,
+                "b": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+                "c": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+                "d": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+            },
+            index=np.random.rand(9 * 3),
+        )
+        psdf = ps.from_pandas(pdf)
+
+        self.assert_eq(
+            psdf.groupby(["a"])["b"].nlargest(1).sort_values(),
+            pdf.groupby(["a"])["b"].nlargest(1).sort_values(),
+        )
+        self.assert_eq(
+            psdf.groupby(["a"])["b"].nlargest(2).sort_index(),
+            pdf.groupby(["a"])["b"].nlargest(2).sort_index(),
+        )
+        self.assert_eq(
+            (psdf.b * 10).groupby(psdf.a).nlargest(2).sort_index(),
+            (pdf.b * 10).groupby(pdf.a).nlargest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.rename().groupby(psdf.a).nlargest(2).sort_index(),
+            pdf.b.rename().groupby(pdf.a).nlargest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.groupby(psdf.a.rename()).nlargest(2).sort_index(),
+            pdf.b.groupby(pdf.a.rename()).nlargest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.rename().groupby(psdf.a.rename()).nlargest(2).sort_index(),
+            pdf.b.rename().groupby(pdf.a.rename()).nlargest(2).sort_index(),
+        )
+        with self.assertRaisesRegex(ValueError, "nlargest do not support multi-index now"):
+            psdf.set_index(["a", "b"]).groupby(["c"])["d"].nlargest(1)
+
+    def test_nsmallest(self):
+        pdf = pd.DataFrame(
+            {
+                "a": [1, 1, 1, 2, 2, 2, 3, 3, 3] * 3,
+                "b": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+                "c": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+                "d": [1, 2, 2, 2, 3, 3, 3, 4, 4] * 3,
+            },
+            index=np.random.rand(9 * 3),
+        )
+        psdf = ps.from_pandas(pdf)
+
+        self.assert_eq(
+            psdf.groupby(["a"])["b"].nsmallest(1).sort_values(),
+            pdf.groupby(["a"])["b"].nsmallest(1).sort_values(),
+        )
+        self.assert_eq(
+            psdf.groupby(["a"])["b"].nsmallest(2).sort_index(),
+            pdf.groupby(["a"])["b"].nsmallest(2).sort_index(),
+        )
+        self.assert_eq(
+            (psdf.b * 10).groupby(psdf.a).nsmallest(2).sort_index(),
+            (pdf.b * 10).groupby(pdf.a).nsmallest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.rename().groupby(psdf.a).nsmallest(2).sort_index(),
+            pdf.b.rename().groupby(pdf.a).nsmallest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.groupby(psdf.a.rename()).nsmallest(2).sort_index(),
+            pdf.b.groupby(pdf.a.rename()).nsmallest(2).sort_index(),
+        )
+        self.assert_eq(
+            psdf.b.rename().groupby(psdf.a.rename()).nsmallest(2).sort_index(),
+            pdf.b.rename().groupby(pdf.a.rename()).nsmallest(2).sort_index(),
+        )
+        with self.assertRaisesRegex(ValueError, "nsmallest do not support multi-index now"):
+            psdf.set_index(["a", "b"]).groupby(["c"])["d"].nsmallest(1)
+
+
+class NlargestNsmallestTests(NlargestNsmallestTestsMixin, PandasOnSparkTestCase):
+    pass
+
+
+if __name__ == "__main__":
+    from pyspark.pandas.tests.groupby.test_nlargest_nsmallest import *  # noqa: F401
+
+    try:
+        import xmlrunner
+
+        testRunner = xmlrunner.XMLTestRunner(output="target/test-reports", verbosity=2)
+    except ImportError:
+        testRunner = None
+    unittest.main(testRunner=testRunner, verbosity=2)
